Begin CRUD transition & fix listingItems collection to all lowercase
