+++ package.json
@@ -8,7 +8,7 @@
     "node": ">=6.13.0"
   },
   "scripts": {
-    "react-dev": "webpack -d --watch",
+    "react": "webpack -d --watch",
     "start": "nodemon --inspect server/index.js",
     "seed": "node database/seed.js",
     "seed-db": "mysql -u root -p < schema.sql",
